faster extra

+++ src/force/potential.cuh
@@ -24,7 +24,7 @@ public:
   // size of the B vector (for each atom) in extrapolation grade calculation
   int B_projection_size = 0;
   // this points to GPU
-  double* B_projection = nullptr;
+  float* B_projection = nullptr;
   bool need_B_projection = false;
 
   int N1;
